Fix a couple of problems with the MotionManager
Problem 1: Many phones are not calibrated properly, and although
assuming that a phone that lies perfectly still returns a total
acceleration of 9.8m/s^2 does not always work. This change compares the
acceleration value with the previously measured value instead.

Problem 2: The previous code stated that it would measure if the
acceleration changed by +/- 0.4g, but instead it measured if the
acceleration changed by +/- 0.4m/s^2, which is a much smaller number.
I've changed this so that the threshold is really 0.4g (~= 3.9m/s^2).