Use a significant motion sensor if possible
Change the MotionManager class to use a significant motion sensor if
available, or fall back to the accelerometer implementation if not.
Using the significant motion sensor should reduce battery usage some.

A timeout task has been introduced and is now used to send the motion
timeout event by both the significant motion implementation and the
accelerometer implementation.