Fixed issue with FAST_SCAN not working on Android 5.1.
In Android 5.1 the shortes interval for a repeating alarm is 60
seconds. The FAST_SCAN uses an alarm with delay of 1667 ms which works
on earlier versions, but in 5.1 it will be extended to 60 seconds. The
workaround is to use a scheduled task instead of alarm manager when it
is needed.

The workaround will not work if device enters deep sleep mode, but
then the compat library should change to SLOW_SCAN and scan period is
more than 60 seconds (thus it will use alarm manager).

See:
http://code.google.com/p/android/issues/detail?id=161244